+++ scripts/ncu.js
@@ -4,7 +4,8 @@ const packageFiles = [
   './src/template/project/package.json',
   './src/template/components/app/package.json',
   './src/template/components/auth/package.json',
-  './src/template/components/provider/package.json'
+  './src/template/components/provider/package.json',
+  './src/template/components/output/package.json'
 ]
 
 async function upgrade () {