+++ src/utils/exec-realtime.js
@@ -5,7 +5,7 @@ const execa = require('execa')
  * @param  {string} command command to execute
  */
 module.exports = (command) => {
-  const result = execa.shell(command)
+  const result = execa(command, { shell: true })
   result.stdout.pipe(process.stdout)
   result.stderr.pipe(process.stderr)
 }